physical-ai-textbook/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==3.4.0
qdrant-client==1.6.2
python-dotenv==1.0.0
openai==1.3.7
google-generativeai==0.8.4
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
asyncpg==0.29.0